Extract remote deck import planning
